api/database/models.py
import os

from dotenv import load_dotenv

from sqlalchemy import Column, Integer, String, DateTime, func
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine


load_dotenv()
Base = declarative_base()


class User(Base):
    __tablename__ = "users"

    id = Column(Integer, primary_key=True)
    created_at = Column(DateTime, default=func.now())
    telegram_id = Column(Integer, unique=True, index=True, nullable=False)
    username = Column(String, nullable=True)
    first_name = Column(String, nullable=True)
    last_name = Column(String, nullable=True)


engine = create_async_engine(os.getenv("SQLALCHEMY_DATABASE_URL"))
session = sessionmaker(bind=engine, class_=AsyncSession, expire_on_commit=False)
