requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.5.0
gradio_client==0.16.4
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
openai==1.30.5
outcome==1.3.0.post0
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
rfc3986==1.5.0
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
websockets==11.0.3
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
